requirements.txt
beautifulsoup4
django-debug-toolbar
psycopg2
scrapy
scrapy-djangoitem